+++ src/plugins/fetch-plugin.ts
@@ -1,0 +1,50 @@
+import axios from "axios";
+import * as esbuild from "esbuild-wasm";
+import localForage from "localforage";
+
+const fileCache = localForage.createInstance({
+  name: "filecache",
+});
+
+export const fetchPlugin = (inputCode: string) => {
+  return {
+    name: "fetch-plugin",
+    setup(build: esbuild.PluginBuild) {
+      build.onLoad({ filter: /.*/ }, async (args: any) => {
+        if (args.path === "index.js") {
+          return {
+            loader: "jsx",
+            contents: inputCode,
+          };
+        }
+        const cachedResult = await fileCache.getItem<esbuild.OnLoadResult>(
+          args.path,
+        );
+        if (cachedResult) {
+          return cachedResult;
+        }
+        const fileType = args.path.match(/.css$/) ? "css" : "jsx";
+        const { data, request } = await axios.get(args.path);
+        const escaped = data
+          .replace(/\n/g, "")
+          .replace(/"/g, '\\"')
+          .replace(/'/g, "\\'");
+        const contents =
+          fileType === "css"
+            ? `
+        const style = document.createElement('style');
+        style.innerText = '${escaped}';
+        document.head.appendChild(style);
+        `
+            : data;
+        const result: esbuild.OnLoadResult = {
+          loader: "jsx",
+          contents: contents,
+          resolveDir: new URL("./", request.responseURL).pathname,
+        };
+        await fileCache.setItem(args.path, result);
+        return result;
+      });
+    },
+  };
+};
