+++ src/plugins/unpkg-path-plugin.ts
@@ -1,57 +1,29 @@
-import axios from "axios";
 import * as esbuild from "esbuild-wasm";
-import localForage from "localforage";
 
-const fileCache = localForage.createInstance({
-  name: "filecache",
-});
-
-export const unpkgPathPlugin = (inputCode: string) => {
+export const unpkgPathPlugin = () => {
   return {
     name: "unpkg-path-plugin",
     setup(build: esbuild.PluginBuild) {
+      // root entry of index.js
+      build.onResolve({ filter: /(^index\.js$)/ }, (args: any) => {
+        return { path: args.path, namespace: "a" };
+      });
+
+      // relative paths in module
+      build.onResolve({ filter: /^\.+\// }, (args: any) => {
+        return {
+          namespace: "a",
+          path: new URL(args.path, "https://unpkg.com" + args.resolveDir + "/")
+            .href,
+        };
+      });
+
+      // main file of a module
       build.onResolve({ filter: /.*/ }, async (args: any) => {
-        if (args.path === "index.js") {
-          return { path: args.path, namespace: "a" };
-        }
-
-        if (args.path.includes("./") || args.path.includes("../")) {
-          return {
-            namespace: "a",
-            path: new URL(
-              args.path,
-              "https://unpkg.com" + args.resolveDir + "/",
-            ).href,
-          };
-        }
-
         return {
           namespace: "a",
           path: `https://unpkg.com/${args.path}`,
         };
-      });
-
-      build.onLoad({ filter: /.*/ }, async (args: any) => {
-        if (args.path === "index.js") {
-          return {
-            loader: "jsx",
-            contents: inputCode,
-          };
-        }
-        const cachedResult = await fileCache.getItem<esbuild.OnLoadResult>(
-          args.path,
-        );
-        if (cachedResult) {
-          return cachedResult;
-        }
-        const { data, request } = await axios.get(args.path);
-        const result: esbuild.OnLoadResult = {
-          loader: "jsx",
-          contents: data,
-          resolveDir: new URL("./", request.responseURL).pathname,
-        };
-        await fileCache.setItem(args.path, result);
-        return result;
       });
     },
   };
